requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycodestyle==2.9.1
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
toml==0.10.2
